changed to real(8), debugging carlos overlap

+++ src/accumalators.f90
@@ -3,7 +3,7 @@ module Accumalators
 
    type :: Accumalator
       integer :: nMoves, nTrans, nRot, nFlip, nVol, nAccept, nTransAccept, nRotAccept, nFlipAccept, nVolAccept
-      real :: ratioTrans, ratioRot, ratioFlip, ratioVol, ratio
+      real(8) :: ratioTrans, ratioRot, ratioFlip, ratioVol, ratio
    end type Accumalator
 
 contains
